fix float in ppg

+++ appgui/PpgData.py
@@ -18,7 +18,7 @@ class PpgData(threading.Thread):
             print("🔌 Connected to client")
             try:
                 async for message in websocket:
-                    self.data_queue.put(int(message))  # oryginalne działanie
+                    self.data_queue.put(float(message))  # oryginalne działanie
             except websockets.exceptions.ConnectionClosed:
                 print("❌ Connection closed")
